Merge (squashed) of the 'docs' branch.
* Main sphinx/pydoc files

* Script to populate the RST files and generate the HTMLs

* Basic documentation files

* The processor file section still need to be written (the files are 'empty')

* The module files are automatic generated, but the pydocs are still to
be written in the python files.

The processor file section is actually dependent of the website map we
come up with, since this is a mix between API documentation and user
documentation.

We can add a README to the docs folder but basically the way to update
the docs is to run the /docs/update-docs.sh script and send the
/docs/_build/html folder to the webserver.

For example:
rsync -rvu --exclude=doctrees $HOME/blender-vr/blender-vr/docs/_build/html/ user@domain:/var/www/blendervr/api/

Note: the generated .rst files are intentionally left out of the git
server (see .gitignore).