Doc: create dummy modules to remove warning from apidocs
This is a hack, big time!
In some cases this can be avoided by simply removing the dependencies
from a top-level call (i.e., to import bge from within a function).

In other cases, however, this is the only way of getting Sphinx to work
properly (for example, when a top-level class inherits a type from
PySide).

I learned of this method in the Blender Conference 2014 while talking to
Campbell Barton. I don't even know if this is the recommended way of
doing it, but it works.