Code/Doc: Add the missing main() functions
To make the docs to compile (warnings free) at the moment you need to
remove the call for main() that follows the # XXX

This need to be refactored in a way of another.
We either stablish that all the top package have a main() function that
is called, or we get rid of the top-level calls entirely.

Those ones are particularly problematic since they invoke a sys.exit()
and sphinx refuses to even import the module in those conditions.