Auto apply changes to configuration
This is a hack to automatically apply changes when: 
1. A new configuration file is selected (so it is not necessary to push "Load configuration" button)
2. A new Screen set is selected (so you don't need to push "Load screen set" button)

This makes the behaviour more intuitive, because the screen setting that you see is always the same than the screen setting that is applied. The "Load configuration" and  "Load screen set" buttons seem to be no longer necessary, but I kept there to help in case of debugging the application.